Roberto's model
This is my Simulink model, which is designed to simulate the charge and discharge behavior of a lithium-ion battery using a Single Particle Model (SPM). Our task was to implement and test various custom charging and discharging profiles that include safety cutoffs based on voltage thresholds, state of charge (SOC), and temperature. This whole model operates as a closed feedback loop to charge and discharge the battery.