SPM Model file
This Simulink model is designed to simulate the charge and discharge behavior of a lithium-ion battery using a Single Particle Model (SPM); In contrast to the group, my model utilizes the SPM battery block on Simscape. The objective of this is to implement and test the charging and discharging algorithm that includes multiple safety cutoffs based on battery's active state, such as state of charge (SOC), voltage, and temperature. This model utilizes a closed feedback loop which circulates real time data as it simulates.